requirements.txt
streamlit>=1.44,<2
PyYAML>=6.0
pytest>=8.0
openai>=1.30.0
python-dotenv>=1.0
pandas>=2.0
altair>=5.0
